+++ Documentation/Changelog.md
@@ -10,6 +10,10 @@ To upgrade from TDW v1.11 to v1.12, read [this guide](upgrade_guides/v1.11_to_v1
 
 - Fixed: `send_albedo_colors` doesn't work.
 
+### `tdw` module
+
+- Added: `type_aliases.PATH`, a type alias for `Union[str, Path]`. Replaced `Union[str, Path]` with `PATH` through TDW's codebase.
+
 ## v1.12.18
 
 ### Command API